moved reports to separate components and added a reports hub page
