+++ tdei-ui/src/routes/Reports/Reports.module.css
@@ -46,3 +46,67 @@
   justify-content: flex-start;
   margin-top: 12px;
 }
+.accordionStack {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+}
+
+.accordion {
+  border: 1px solid #e9eef5 !important;
+  border-radius: 12px !important;
+  overflow: hidden;
+  box-shadow: 0 2px 10px rgba(16,24,40,0.04);
+  background: #fff;
+}
+
+.accSummary {
+  background: #f9fafb;
+  padding: 8px 16px !important;
+}
+
+.accHeader {
+  display: flex;
+  flex-direction: column;
+  gap: 2px;
+}
+
+.accTitle {
+  font-weight: 700 !important;
+  color: #0f172a;
+}
+
+.accSubtitle {
+  font-size: 13px !important;
+  color: #6b7280;
+}
+
+.accDetails {
+  background: #fff;
+  padding: 14px 16px 16px !important;
+}
+
+/* Users row */
+.usersRow {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.usersText {
+  color: #374151;
+  font-size: 14px;
+}
+.downloadButton {
+  background: var(--primary-color);
+  border: none;
+  padding: 6px;
+  cursor: pointer;
+  transition: background 0.2s;
+  min-width: 100px;
+}
+
+.downloadButton:hover {
+  background: var(--tdei-blue);
+  border-radius: 4px;
+}
